// src/types/user-scopes.ts
export type UserScopes =
    | "merchant.view"
    | "merchant.update"
    | "product.view"
    | "product.create"
    | "product.update"
    | "inventory.view"
    | "inventory.update"
    | "checkout.create"
    | "checkout.refund"
    | "order.view"
    | "order.update"
    | "customer.view"
    | "customer.create"
    | "staff.view"
    | "staff.manage"
    | "report.view"
    | "settings.update"
    | "subscription.view"
    | "platform.manage";